+++ test/test.js
@@ -25,14 +25,28 @@ var simpleDriverFromPromise = makeAsyncDriver({
     ) 
 })
 
+var driverWithFailure = makeAsyncDriver({
+  normalizeRequest: (name) =>
+    typeof name === 'string' ? {name} : name,
+  getResponse: (request, cb) => {
+    return new Promise((resolve, reject) =>
+      request.name
+        ? resolve('async ' + request.name)
+        : reject('async error')
+    )
+  },
+  responseProp: true
+})
+
+
 var asyncDriver = makeAsyncDriver({
   createResponse$: (request) =>
     O.create(observer => {
       setTimeout(() => observer.onNext({asyncName: 'async ' + request.name}), 10)
     }),
   requestProp: 'query',
-  normalize: (name) =>
-    typeof name ==='string'
+  normalizeRequest: (name) =>
+    typeof name === 'string'
       ? {name: name.toUpperCase()}
       : {...name, name: name.name.toUpperCase()},
   isolateProp: '_scope',
@@ -151,4 +165,80 @@ test('Two isolated requests', (t) => {
       })
     }
   })
+})
+
+test('successful and failed helpers', (t) => {
+  let mergedCount = 0
+  let successfulCount = 0
+  let failedCount = 0
+  const main = ({async}) => {
+    return {
+      async: O.of('John', 'Mary', '', 'Alex', 'Jane', '', 'Mike').delay(1),
+      merged: async.mergeAll().catch(O.empty()),
+      success: async
+        .filter(r$ => r$.request.name)
+        .filter(r$ => r$.request.name !== 'Alex')
+        .successful(({response, request}) =>
+        ({response: response + ' mapped', request})
+      ),
+      fail: async.failed(({error, request}) =>
+        ({error: error + ' mapped', request})
+      )
+    }
+  }
+  run(main, {
+    async: driverWithFailure,
+    merged: (response$) => {
+      var count = 0
+      response$.subscribe(r => {
+        mergedCount++
+      })
+    },
+    success: (response$) => {
+      response$.subscribe(r => {
+        successfulCount++
+        //console.log('success', r)
+        t.is(r.response, `async ${r.request.name} mapped`, 'successful response mapped')
+      })
+    },
+    fail: (error$) => {
+      error$.subscribe(r => {
+        failedCount++
+        t.is(r.error, 'async error mapped', 'error in `error` wrapped ok')
+        t.is(r.request.name, '', 'error contains request')
+      })
+    }
+  })
+  setTimeout(() => {
+    t.is(mergedCount, 2, 'merged count correct')
+    t.is(successfulCount, 4, 'successful count correct')
+    t.is(failedCount, 2, 'failed count correct')
+    t.end()
+  }, 100)
+})
+
+test('select helper', (t) => {
+  let selectedCount = 0
+  const main = ({async}) => {
+    return {
+      async: O.from([
+        {name: 'John', category: 'good'},
+        {name: 'Alex', category: 'bad'},
+        {name: 'Jane', category: 'good'}
+      ]).delay(0),
+      result: async.select('good').successful()
+    }
+  }
+  run(main, {
+    async: simpleDriver,
+    result: (response$) => {
+      response$.subscribe(r => {
+        selectedCount++
+      })
+    }
+  })
+  setTimeout(() => {
+    t.is(selectedCount, 2, 'selected count correct')
+    t.end()
+  }, 100)
 })