+++ HISTORY.md
@@ -1,5 +1,10 @@
+### v1.2.0     
+    * Added flatten helpers `successful` and `failed`
+    * Added default selector helper `select` with `category` prop
+    * Exports `attachHelpers` and other helper methods for extenal use
+
 ### v1.1.2
-    * use `normalizeRequest` in isolate if no `isolateMap`     
+    * Use `normalizeRequest` in isolate if no `isolateMap`     
 
 ### v1.1.1
     * Added ability to create driver repsonse from callback and (native) Promise (`getResponse`)